+++ yahoo-quotes/spring-cloud-starter-stream-source-yahoo-quotes/src/main/java/org/springframework/cloud/stream/app/yahoo/quotes/source/utils/PartitionedList.java
@@ -15,7 +15,7 @@
  *
  */
 
-package org.springframework.cloud.stream.app.yahooquotes.source.utils;
+package org.springframework.cloud.stream.app.yahoo.quotes.source.utils;
 
 import java.util.AbstractList;
 import java.util.List;
