+++ yahoo-quotes/spring-cloud-starter-stream-source-yahoo-quotes/src/main/resources/META-INF/spring-configuration-metadata-whitelist.properties
@@ -15,5 +15,5 @@
 #
 #
 
-configuration-properties.classes=org.springframework.cloud.stream.app.yahooquotes.source.YahooQuotesSourceConfiguration
+configuration-properties.classes=org.springframework.cloud.stream.app.yahoo.quotes.source.YahooQuotesSourceConfiguration
