+++ yahoo-quotes/spring-cloud-starter-stream-source-yahoo-quotes/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+yahoo.quotes.cron-expression="0 * 9-17 * * MON-FRI"
+yahoo.quotes.batch-size=400
+yahoo.quotes.zone=EST
