+++ yahoo-quotes/spring-cloud-starter-stream-source-yahoo-quotes/src/test/java/org/springframework/cloud/stream/app/yahoo/quotes/source/YahooQuotesClientTest.java
@@ -15,7 +15,7 @@
  *
  */
 
-package org.springframework.cloud.stream.app.yahooquotes.source;
+package org.springframework.cloud.stream.app.yahoo.quotes.source;
 
 import java.util.Arrays;
 import java.util.Collections;
